Handles `noarch_python: true` and `noarch: python`
If a recipe defines itself to be a noarch Python package either using
the legacy definition or the new one, make sure to pop it from the spec
before getting into matrix resolution. This is what is also done with
NumPy when `x.x` is used. So this seems to be a very sound way to solve
the problem. Also when checking all of these settings, make sure to
define a default value if the parameter is not found in the `meta`.