Support `noarch`/`noarch_python` in `DummyPackage`
The `DummyPackage` is an internal type used for testing only. To be able
to just run existing unit tests with `noarch` or `noarch_python`, we
need to make sure the `DummyPackage` can handle these cases somehow.
Hence we set these attributes to their default values, but allow them to
be overridden. Also we return them on a `get_value` call if the user
wants them.